requirements.txt
fastapi
hypercorn
sentence-transformers
faiss-cpu
numpy
uvloop
httptools 
